+++ R/plot_scan1_heatmap.R
@@ -52,7 +52,7 @@
 #' plot_scan1_heatmap(out, map)
 #'
 #' # make the y-axis quantitative
-#' y <- as.numeric(grav2$phenocovar$time)
+#' y <- grav2$phenocovar$time
 #  plot_scan1_heatmap(out, map, y=y, ylab=colnames(grav2$phenocovar))
 #'
 #' # plot with colors white -> darkslateblue, and with color scale